requirements.txt
requests==2.31.0
Jinja2==3.1.2
python-dotenv==1.1.1
pydantic==2.11.7
qrcode==8.2